Move all session related logic to one package
