+++ backend/t4cclient/sessions/database.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from sqlalchemy.orm import Session
-from t4cclient.sql_models.sessions import DatabaseSession
+from t4cclient.sessions.models import DatabaseSession
 
 
 def get_sessions_for_user(db: Session, username: str):
