+++ backend/t4cclient/sessions/garbage_cleaner.py
@@ -5,8 +5,8 @@ import logging
 
 import click
 from t4cclient.core import database
-from t4cclient.core.database import sessions
-from t4cclient.routes.sessions import inject_attrs_in_sessions
+from t4cclient.sessions import database
+from t4cclient.sessions.routes import inject_attrs_in_sessions
 
 LOGGER = logging.getLogger(__name__)
 
@@ -14,10 +14,10 @@ LOGGER = logging.getLogger(__name__)
 @click.command()
 def clean_old_sessions():
     with database.SessionLocal() as db:
-        for session in inject_attrs_in_sessions(sessions.get_all_sessions(db)):
+        for session in inject_attrs_in_sessions(database.get_all_sessions(db)):
             if session["state"] == "exited":
                 try:
-                    sessions.delete_session(db, session["id"])
+                    database.delete_session(db, session["id"])
                 except Exception as e:
                     LOGGER.info(
                         "Removing then container with id %s failed",
