+++ backend/t4cclient/sessions/sessions.py
@@ -14,7 +14,7 @@ import requests
 from t4cclient import config
 from t4cclient.core.operators import OPERATOR
 from t4cclient.schemas.sessions import WorkspaceType
-from t4cclient.sql_models.sessions import DatabaseSession
+from t4cclient.sessions.models import DatabaseSession
 
 log = logging.getLogger(__name__)
