+++ backend/t4cclient/sql_models/__init__.py
@@ -2,4 +2,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # These import statements of the models are required and should not be removed! (SQLAlchemy will not load the models otherwise)
-from . import notices, repositories, sessions, users
+from ..sessions import models
+from . import notices, repositories, users
